pyproject.toml
[project]
name = "mariorl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"

# --- Dependencies ---

dependencies = [
    "gymnasium>=1.2.0",
    "stable-baselines3[extra]>=2.7.0",
    "stable-retro>=0.9.5",
]

# --- Sources ---

[[tool.uv.index]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu126"

[tool.uv.sources]
torch = { index = "torch-gpu" }
torchaudio = { index = "torch-gpu" }
torchvision = { index = "torch-gpu" }

# --- USEFUL COMMANDS ---

# source .venv/bin/activate

# use the below command to delete .venv and uv.lock for clean start 👇

# Remove-Item -Recurse -Force .venv
# Remove-Item -Force uv.lock